Update to pull #38
Had some time before I left for the day so I decided to improve this with the fix discussed earlier.

off-click-filter is defined as `off-click-filter="#test-id,.test-class"` and is applied to elements that are filters of the offclick on the element with that id or class. Meaning, an element can still be an filter to multiple off-click elements.

ES6, so you'll have to put babel in here or some other transpiler.